requirements.txt
streamlit
langchain
langchain-community
langchain-openai
duckdb
pandas
chromadb
faiss-cpu
sentence-transformers
plotly
tavily-python
jinja2
pyyaml
python-dotenv
umap-learn
scikit-learn
networkx
pyvis
scipy
seaborn
